+++ rpc/system/internal/model/user.go
@@ -9,11 +9,11 @@ type User struct {
 	Username string `gorm:"type:varchar(15);not null;uniqueIndex:idx_unique_username;comment:学号" json:"username"`
 	Email    string `gorm:"type:varchar(255);uniqueIndex:idx_unique_email;comment:邮箱" json:"email"`
 	Avatar   string `gorm:"type:varchar(255);comment:头像" json:"avatar"`
-	Role     uint   `gorm:"not null;comment:当前角色" json:"role"`
-	Status   int8   `gorm:"not null;default:1;comment:状态(1:正常, 2:禁用)" json:"status"`
+	Role     uint64 `gorm:"not null;comment:当前角色" json:"role"`
+	Status   int32   `gorm:"not null;default:1;comment:状态(1:正常, 2:禁用)" json:"status"`
 	Nickname string `gorm:"type:varchar(255);comment:昵称" json:"nickname"`
 	Phone    string `gorm:"type:char(11);uniqueIndex:idx_unique_phone;comment:手机号" json:"phone"`
-	Gender   int8   `gorm:"not null;default:1;comment:性别(1:男, 2:女)" json:"gender"`
+	Gender   int32   `gorm:"not null;default:1;comment:性别(1:男, 2:女)" json:"gender"`
 	Major    string `gorm:"type:varchar(255);not null;comment:专业" json:"major"`
 	College  string `gorm:"type:varchar(255);not null;comment:学院" json:"college"`
 	Grade    string `gorm:"type:char(4);not null;comment:年级" json:"grade"`
